# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Piteka Accountability API
  description: >
    REST API for the Piteka enterprise accountability workbench.
    First-slice endpoints cover the action-request lifecycle:
    propose, approve, reject, and revoke.
  version: 0.1.0
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
  /api/v1/action-requests:
    get:
      tags:
        - action-requests
      summary: List all action requests
      description: Returns a compact summary of all action requests.
      operationId: listActionRequests
      responses:
        "200":
          description: List of action request summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActionRequestSummaryDtoV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"

    post:
      tags:
        - action-requests
      summary: Propose a new action request
      description: >
        Creates a new action request in `Pending` status.
        Supports idempotency via the `Idempotency-Key` header.
      operationId: proposeActionRequest
      headers:
        Idempotency-Key:
          schema:
            type: string
          description: Idempotency key for safe retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionRequestRequestV1"
      responses:
        "201":
          description: Action request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequestResponseV1"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "409":
          description: Idempotency key conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/action-requests/{id}:
    get:
      tags:
        - action-requests
      summary: Get an action request
      description: Returns full detail for a single action request.
      operationId: getActionRequest
      parameters:
        - name: id
          in: path
          required: true
          description: Action request identifier
          schema:
            type: string
      responses:
        "200":
          description: Action request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequestResponseV1"
        "404":
          description: Action request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/action-requests/{id}/approve:
    post:
      tags:
        - action-requests
      summary: Approve an action request
      description: >
        Approves a pending action request. Requires the approver's user id,
        the intent digest reviewed, and the expected version for CAS.
        Supports idempotency via the `Idempotency-Key` header.
      operationId: approveActionRequest
      parameters:
        - name: id
          in: path
          required: true
          description: Action request identifier
          schema:
            type: string
      headers:
        Idempotency-Key:
          schema:
            type: string
          description: Idempotency key for safe retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveActionRequestRequestV1"
      responses:
        "200":
          description: Action request approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequestResponseV1"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "404":
          description: Action request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "409":
          description: "Conflict: invalid state or version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/action-requests/{id}/reject:
    post:
      tags:
        - action-requests
      summary: Reject an action request
      description: >
        Rejects a pending action request. Requires the approver's user id,
        the intent digest reviewed, and the expected version for CAS.
        Supports idempotency via the `Idempotency-Key` header.
      operationId: rejectActionRequest
      parameters:
        - name: id
          in: path
          required: true
          description: Action request identifier
          schema:
            type: string
      headers:
        Idempotency-Key:
          schema:
            type: string
          description: Idempotency key for safe retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectActionRequestRequestV1"
      responses:
        "200":
          description: Action request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequestResponseV1"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "404":
          description: Action request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "409":
          description: "Conflict: invalid state or version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/action-requests/{id}/revoke:
    post:
      tags:
        - action-requests
      summary: Revoke an approved action request
      description: >
        Revokes an approved action request before dispatch.
        Requires the approver's user id and the expected version for CAS.
        Supports idempotency via the `Idempotency-Key` header.
      operationId: revokeActionRequest
      parameters:
        - name: id
          in: path
          required: true
          description: Action request identifier
          schema:
            type: string
      headers:
        Idempotency-Key:
          schema:
            type: string
          description: Idempotency key for safe retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeActionRequestRequestV1"
      responses:
        "200":
          description: Action request revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionRequestResponseV1"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "404":
          description: Action request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "409":
          description: "Conflict: invalid state or version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/receipts:
    get:
      tags:
        - read-model
      summary: List receipt projections, newest first
      operationId: listReceipts
      responses:
        "200":
          description: Receipt summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReceiptSummaryDtoV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/receipts/{id}:
    get:
      tags:
        - read-model
      summary: One receipt projection with its evidence references
      operationId: getReceipt
      parameters:
        - name: id
          in: path
          required: true
          description: Lower-case hex object identifier
          schema:
            type: string
      responses:
        "200":
          description: The receipt projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResponseV1"
        "404":
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
  /api/v1/receipts/{id}/export:
    get:
      tags:
        - read-model
      summary: Export the receipt's evidence manifest
      description: >
        Returns the content-addressed export manifest for offline verification by
        the Parwana CLI. The manifest is opaque to this API: its shape is defined
        by the pinned Parwana contract, not by Piteka.
      operationId: exportReceipt
      parameters:
        - name: id
          in: path
          required: true
          description: Lower-case hex object identifier
          schema:
            type: string
      responses:
        "200":
          description: Export manifest bytes
          content:
            application/json:
              schema:
                type: object
                description: Parwana-defined export manifest; not a Piteka-owned shape.
                additionalProperties: true
        "404":
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseV1"
...
components:
  schemas:
...
    ActionRequestSummaryDtoV1:
      type: object
      required:
        - id
        - requested_by
        - status
        - created_at
      properties:
        id:
          type: string
          description: The request identifier.
        requested_by:
          type: string
          description: The user who proposed the request.
        status:
          $ref: "#/components/schemas/ActionRequestStatusDtoV1"
        created_at:
          type: integer
          format: int64
          description: Creation time, Unix seconds.
...
    CreateActionRequestRequestV1:
      type: object
      required:
        - requested_by
      properties:
        requested_by:
          type: string
          description: The user proposing the action.
        intent_id:
          type: string
          description: Parwana intent digest (lower-case hex), if already constructed.
    ApproveActionRequestRequestV1:
      type: object
      required:
        - approver_id
        - version
      properties:
        approver_id:
          type: string
          description: The approver's user id.
        intent_id:
          type: string
          description: The intent digest the approver reviewed.
        version:
          type: integer
          format: int64
          description: The expected version for optimistic concurrency (CAS).
...
    ErrorResponseV1:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetailDtoV1"
...